requirements.txt
aiogram
python-dotenv==0.18.0
aiohttp
requests
emoji